♻️ Refactor Account page to server component
