+++ src/locales/ko/user.json
@@ -1,6 +1,7 @@
 {
   "login": "로그인",
   "join": "회원가입",
+  "info": "회원정보",
   "form": {
     "email": "이메일",
     "nickname": "닉네임",
@@ -12,6 +13,9 @@
     "login": "로그인",
     "join": "회원가입",
     "findPassword": "비밀번호 찾기",
-    "joinUs": "회원 가입하기"
+    "joinUs": "회원 가입하기",
+    "logout": "로그아웃",
+    "unsubscribe": "회원탈퇴",
+    "edit": "정보수정"
   }
 }
